Migrate dataset creation/splitting to asapdiscovery.ml.utils.
